Tidy up code results outputting:
Change the code to output all scan results to a `scans/` dir that is
marked by .gitignore.

The code produces a scanned image, scanned pdf and the text content of
the image into that dir using the same name as the input image.

Tidy up some of the CLI setup while achieving this.